gcp/requirements.txt
numpy>=1.18.5
tensorflow
Pillow==9.0.0
Werkzeug==2.2.2
google-cloud-storage==2.11.0